Move DiagramContinuousState into its own file.
