+++ drake/systems/framework/diagram_continuous_state.cc
@@ -1,0 +1,10 @@
+#include "drake/systems/framework/diagram_continuous_state.h"
+
+namespace drake {
+namespace systems {
+
+template class DiagramContinuousState<double>;
+
+}  // namespace systems
+}  // namespace drake
+
